+++ docs/js/main.js
@@ -8,6 +8,11 @@ import(componentUrl).then(res => {
 
   const $console = document.getElementById('console');
 
+  document.addEventListener('clipboard-copy-success', evt => {
+    console.log('clipboard-copy-success ->', evt.detail);
+    $console.innerHTML += `<div>$ <span class="success">clipboard-copy-success</span> -> ${JSON.stringify(evt.detail)}</div>`;
+  });
+
   document.addEventListener('clipboard-copy-error', evt => {
     console.log('clipboard-copy-error ->', evt.detail);
     $console.innerHTML += `<div>$ <span class="error">clipboard-copy-error</span> -> ${evt.detail.error.name}: ${evt.detail.error.message}</div>`;