+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking Changes
 
 - Remove `clipboard-copy:click` event.
+- Remove `button` slot.
 - Rename `clipboard-copy:success` event to `clipboard-copy-success`.
 - Rename `clipboard-copy:error` event to `clipboard-copy-error`.
 - Rename `button-content` slot to `copy`.